Got account update working
OK.  Some props take a longer time to load, so they are not available right away ad appear NULL in componentDidLoad.  The way I solved this ones 1) Wrap the component (e.g. a form in my case) inside a parent component.  This was a good idea in this case anyway.  Then -- in the parent compoent -- you check the prop/state in the render method and return null if it is not set.  Once that state/prop is set, the component will render itself and the child component with all the props readily available.